refactor: separate context creation in core/context
